+++ lib/features/dogs/domain/repositories/dog_repository.dart
@@ -1,8 +1,9 @@
+import 'package:clean_architecture_sample/core/result/result.dart';
 import 'package:clean_architecture_sample/features/dogs/domain/entities/breed.dart';
 import 'package:clean_architecture_sample/features/dogs/domain/entities/dog_image.dart';
 
 abstract class DogRepository {
-  Future<List<Breed>> getBreeds();
-  Future<DogImage> getRandomImage();
-  Future<DogImage> getBreedRandomImage(String breed);
+  Future<Result<List<Breed>>> getBreeds();
+  Future<Result<DogImage>> getRandomImage();
+  Future<Result<DogImage>> getBreedRandomImage(String breed);
 }