+++ lib/features/dogs/domain/usecases/get_breed_random_image.dart
@@ -1,3 +1,4 @@
+import 'package:clean_architecture_sample/core/result/result.dart';
 import 'package:clean_architecture_sample/features/dogs/domain/entities/dog_image.dart';
 import 'package:clean_architecture_sample/features/dogs/domain/repositories/dog_repository.dart';
 
@@ -5,5 +6,5 @@ class GetBreedRandomImage {
   final DogRepository repo;
 
   GetBreedRandomImage(this.repo);
-  Future<DogImage> call(String breed) => repo.getBreedRandomImage(breed);
+  Future<Result<DogImage>> call(String breed) => repo.getBreedRandomImage(breed);
 }