+++ lib/features/dogs/domain/usecases/get_breeds.dart
@@ -1,3 +1,4 @@
+import 'package:clean_architecture_sample/core/result/result.dart';
 import 'package:clean_architecture_sample/features/dogs/domain/entities/breed.dart';
 import 'package:clean_architecture_sample/features/dogs/domain/repositories/dog_repository.dart';
 
@@ -5,5 +6,5 @@ class GetBreeds {
   final DogRepository repo;
 
   GetBreeds(this.repo);
-  Future<List<Breed>> call() => repo.getBreeds();
+  Future<Result<List<Breed>>> call() => repo.getBreeds();
 }