+++ lib/features/dogs/domain/usecases/get_random_image.dart
@@ -1,3 +1,4 @@
+import 'package:clean_architecture_sample/core/result/result.dart';
 import 'package:clean_architecture_sample/features/dogs/domain/entities/dog_image.dart';
 import 'package:clean_architecture_sample/features/dogs/domain/repositories/dog_repository.dart';
 
@@ -5,5 +6,5 @@ class GetRandomImage {
   final DogRepository repo;
 
   GetRandomImage(this.repo);
-  Future<DogImage> call() => repo.getRandomImage();
+  Future<Result<DogImage>> call() => repo.getRandomImage();
 }