+++ lib/features/dogs/presentation/pages/breeds_page.dart
@@ -1,4 +1,5 @@
 import 'package:clean_architecture_sample/core/extensions/async_value_extensions.dart';
+import 'package:clean_architecture_sample/core/extensions/result_extensions.dart';
 import 'package:clean_architecture_sample/features/dogs/routes.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/src/widgets/framework.dart';
@@ -17,22 +18,24 @@ class BreedsPage extends ConsumerWidget {
     return Scaffold(
       appBar: AppBar(title: const Text('Dog Breeds'),),
       body: async.whenOrDefault(
-          data: (breeds) => ListView.separated(
-              itemBuilder: (_, i) {
-                final b = breeds[i];
-                final title = b.subBreeds.isEmpty
-                  ? b.name : '${b.name} (${b.subBreeds.length})';
+          data: (res) => res.toWidget(
+            ok: (breeds) => ListView.separated(
+                itemBuilder: (_, i) {
+                  final b = breeds[i];
+                  final title = b.subBreeds.isEmpty
+                      ? b.name : '${b.name} (${b.subBreeds.length})';
 
-                return ListTile(
-                  title: Text(title),
-                  subtitle: b.subBreeds.isEmpty ? null : Text(b.subBreeds.join(', ')),
-                  onTap: () => context.push('${DogsRoutes.breedRandom}/${b.name}'),
-                );
-              },
-              separatorBuilder: (_, _) {
-                return const Divider(height: 1,);
-              },
-              itemCount: breeds.length
+                  return ListTile(
+                    title: Text(title),
+                    subtitle: b.subBreeds.isEmpty ? null : Text(b.subBreeds.join(', ')),
+                    onTap: () => context.push('${DogsRoutes.breedRandom}/${b.name}'),
+                  );
+                },
+                separatorBuilder: (_, _) {
+                  return const Divider(height: 1,);
+                },
+                itemCount: breeds.length
+            )
           ),
       ),
       floatingActionButton: FloatingActionButton.extended(